fix(docker-entrypoint): do not clobber existing config for defaults

+++ docker-entrypoint.sh
@@ -20,13 +20,15 @@ ARTIFACT_URL="${RELEASES_URL}/download/${RELEASE_TAG}/VanillaVotifier.jar"
 >&2 echo "[info] Downloading ${ARTIFACT_URL}..."
 curl -fsSL "${ARTIFACT_URL}" -o /data/VanillaVotifier.jar
 
-if [[ ! $CONFIG_YAML_URL ]]; then
-  >&2 echo "[warn] \$CONFIG_YAML_URL not specified, using default"
-  CONFIG_YAML_URL="https://raw.githubusercontent.com/xMamo/VanillaVotifier/${RELEASE_TAG}/src/main/resources/mamo/vanillaVotifier/config.yaml"
-fi
+if [[ ! $CONFIG_YAML_URL && ! -f /data/config.yaml ]]; then
+  if [[ ! $CONFIG_YAML_URL ]]; then
+    >&2 echo "[warn] \$CONFIG_YAML_URL not specified, using default"
+    CONFIG_YAML_URL="https://raw.githubusercontent.com/xMamo/VanillaVotifier/${RELEASE_TAG}/src/main/resources/mamo/vanillaVotifier/config.yaml"
+  fi
 
->&2 echo "[info] Downloading ${CONFIG_YAML_URL}..."
-curl -fsSL "${CONFIG_YAML_URL}" -o /data/config.yaml
+  >&2 echo "[info] Downloading ${CONFIG_YAML_URL}..."
+  curl -fsSL "${CONFIG_YAML_URL}" -o /data/config.yaml
+fi
 
 if [[ $REPLACE_ENV ]]; then
   >&2 echo "[info] Interpolating config.yaml with CFG variables..."
